requirements.txt
osmnx 
geopandas 
geojson
flask
networkx 
geopandas
flask-cors
scikit-learn
python-dotenv
requests
os